+++ app/Main.hs
@@ -11,7 +11,7 @@ run :: IO ()
 run = do
   putStrLn "Running for 1 minute..."
   runPhilosophers 5
-  threadDelay $ 1000 * 1000 * 1000 * 60 -- 1 minute
+  threadDelay $ 1000 * 1000 * 60 -- 1 minute
 
 main :: IO ()
 main = run `catch` handler