+++ src/Control/Concurrent/STM/Free/Internal/Impl.hs
@@ -1,31 +1,13 @@
 module Control.Concurrent.STM.Free.Internal.Impl where
 
-import           Control.Concurrent.MVar                               (MVar,
-                                                                        newMVar,
-                                                                        putMVar,
+import           Control.Concurrent.MVar                               (putMVar,
                                                                         takeMVar)
-import           Control.Exception                                     (NonTermination,
-                                                                        catch)
-import           Control.Monad.Free
-import           Control.Monad.IO.Class                                (liftIO)
-import           Control.Monad.State.Strict                            (StateT, evalStateT,
-                                                                        get,
-                                                                        modify,
-                                                                        put,
-                                                                        runStateT)
-import           Data.Aeson                                            (FromJSON,
-                                                                        ToJSON,
-                                                                        decode,
-                                                                        encode)
-import qualified Data.Aeson                                            as A
-import qualified Data.ByteString.Lazy                                  as BSL
-import           Data.IORef                                            (IORef, modifyIORef,
-                                                                        newIORef,
-                                                                        readIORef,
-                                                                        writeIORef)
+import           Control.Monad.State.Strict                            (runStateT)
+import           Data.IORef                                            (newIORef,
+                                                                        readIORef)
 import qualified Data.Map                                              as Map
-import           Data.Time.Clock                                       (UTCTime, getCurrentTime)
-import           GHC.Generics                                          (Generic)
+import           Data.Unique                                           (newUnique)
+
 
 import           Control.Concurrent.STM.Free.Internal.Common
 import           Control.Concurrent.STM.Free.Internal.STML.Interpreter
@@ -35,24 +17,24 @@ import           Control.Concurrent.STM.Free.TVar
 
 
 cloneTVarHandle :: (TVarId, TVarHandle) -> IO (TVarId, TVarHandle)
-cloneTVarHandle (tvarId, TVarHandle _ timestamp tvarData) = do
+cloneTVarHandle (tvarId, TVarHandle ustamp tvarData) = do
   newTVarData <- readIORef tvarData >>= newIORef
-  pure (tvarId, TVarHandle tvarId timestamp newTVarData)
+  pure (tvarId, TVarHandle ustamp newTVarData)
 
-takeSnapshot :: Context -> IO (Timestamp, TVars)
+takeSnapshot :: Context -> IO (UStamp, TVars)
 takeSnapshot (Context mtvars) = do
   tvars <- takeMVar mtvars
   tvarKVs <- mapM cloneTVarHandle (Map.toList tvars)
   putMVar mtvars tvars
-  timestamp <- getCurrentTime
-  pure (timestamp, Map.fromList tvarKVs)
+  ustamp <- newUnique
+  pure (ustamp, Map.fromList tvarKVs)
 
-tryCommit :: Context -> Timestamp -> TVars -> IO Bool
-tryCommit (Context mtvars) timestamp stagedTVars = do
+tryCommit :: Context -> UStamp -> TVars -> IO Bool
+tryCommit (Context mtvars) ustamp stagedTVars = do
   tvars <- takeMVar mtvars
 
   let conflict = Map.foldMapWithKey (f tvars "") stagedTVars
-  let newTVars = Map.unionWith (merge timestamp) stagedTVars tvars
+  let newTVars = Map.unionWith (merge ustamp) stagedTVars tvars
 
   putMVar mtvars $ if null conflict then newTVars else tvars
 
@@ -60,22 +42,21 @@ tryCommit (Context mtvars) timestamp stagedTVars = do
 
   where
     f :: TVars -> String -> TVarId -> TVarHandle -> String
-    f origTvars acc tvarId (TVarHandle _ stagedTS _) = case Map.lookup tvarId origTvars of
-      Nothing                                           -> acc
-      Just (TVarHandle _ origTS _) | origTS == stagedTS -> acc
-                                   | otherwise          -> acc ++ " " ++ show tvarId
-    merge :: Timestamp -> TVarHandle -> TVarHandle -> TVarHandle
-    merge ts' (TVarHandle tvarId _ d) _ = TVarHandle tvarId ts' d
-
+    f origTvars acc tvarId (TVarHandle stagedUS _) = case Map.lookup tvarId origTvars of
+      Nothing                                         -> acc
+      Just (TVarHandle origUS _) | origUS == stagedUS -> acc
+                                 | otherwise          -> acc ++ " " ++ show tvarId
+    merge :: UStamp -> TVarHandle -> TVarHandle -> TVarHandle
+    merge us' (TVarHandle _ d) _ = TVarHandle us' d
 
 runSTM :: Int -> Context -> STML a -> IO a
 runSTM delay ctx stml = do
-  (timestamp, snapshot)              <- takeSnapshot ctx
-  (eRes, AtomicRuntime _ stagedTVars) <- runStateT (runSTML stml) (AtomicRuntime timestamp snapshot)
+  (ustamp, snapshot)                  <- takeSnapshot ctx
+  (eRes, AtomicRuntime _ stagedTVars) <- runStateT (runSTML stml) (AtomicRuntime ustamp snapshot)
   case eRes of
     Left RetryCmd -> runSTM (delay * 2) ctx stml      -- TODO: tail recursion
     Right res     -> do
-      success <- tryCommit ctx timestamp stagedTVars
+      success <- tryCommit ctx ustamp stagedTVars
       if success
         then return res
         else runSTM (delay * 2) ctx stml      -- TODO: tail recursion