+++ src/Control/Concurrent/STM/Free/Internal/STML/Interpreter.hs
@@ -1,54 +1,36 @@
-{-# LANGUAGE LambdaCase #-}
 module Control.Concurrent.STM.Free.Internal.STML.Interpreter where
 
-import           Control.Concurrent.MVar                     (MVar, newMVar,
-                                                              putMVar, readMVar,
-                                                              takeMVar,
-                                                              tryReadMVar)
-import           Control.Monad.Free
+import           Control.Monad.Free                          (Free (..))
 import           Control.Monad.IO.Class                      (liftIO)
-import           Control.Monad.State.Strict                  (StateT,
-                                                              evalStateT, get,
-                                                              modify, put)
+import           Control.Monad.State.Strict                  (get, put)
 import           Data.Aeson                                  (FromJSON, ToJSON,
                                                               decode, encode)
-import qualified Data.Aeson                                  as A
-import qualified Data.ByteString.Lazy                        as BSL
-import           Data.IORef                                  (IORef,
-                                                              modifyIORef,
-                                                              newIORef,
-                                                              readIORef,
+import           Data.IORef                                  (IORef, readIORef,
                                                               writeIORef)
 import qualified Data.Map                                    as Map
-import           Data.Time.Clock                             (UTCTime,
-                                                              getCurrentTime)
-import           GHC.Generics                                (Generic)
 
 import           Control.Concurrent.STM.Free.Internal.Common
 import           Control.Concurrent.STM.Free.Internal.Types
 import           Control.Concurrent.STM.Free.STML
 import           Control.Concurrent.STM.Free.TVar
 
--- TODO: reading TVar - at any time is successful operation or it should be aware of TVar changing in other transactions?
-
 newTVar' :: ToJSON a => a -> Atomic (TVar a)
 newTVar' a = do
-  AtomicRuntime timestamp tvars <- get
+  AtomicRuntime ustamp tvars <- get
 
-  let nextId = Map.size tvars
-  tvarHandle <- liftIO $ createTVar timestamp nextId a
+  (tvarId, tvarHandle) <- liftIO $ createTVar ustamp a
 
-  let newTvars = Map.insert nextId tvarHandle tvars
-  put $ AtomicRuntime timestamp newTvars
-  pure $ TVar nextId
+  let newTvars = Map.insert tvarId tvarHandle tvars
+  put $ AtomicRuntime ustamp newTvars
+  pure $ TVar tvarId
 
 readTVar' :: FromJSON a => TVar a -> Atomic a
 readTVar' (TVar tvarId) = do
-  AtomicRuntime timestamp tvars <- get
+  AtomicRuntime ustamp tvars <- get
 
   case Map.lookup tvarId tvars of
     Nothing                        -> error $ "Impossible: TVar not found: " ++ show tvarId
-    Just (TVarHandle _ _ tvarData) -> do
+    Just (TVarHandle _ tvarData) -> do
       s <- liftIO $ readIORef tvarData
       case decode s of
         Nothing -> error $ "Impossible: Decode error of TVar: " ++ show tvarId
@@ -56,11 +38,11 @@ readTVar' (TVar tvarId) = do
 
 writeTVar' ::  ToJSON a => TVar a -> a -> Atomic ()
 writeTVar' (TVar tvarId) a = do
-  AtomicRuntime timestamp tvars <- get
+  AtomicRuntime ustamp tvars <- get
 
   case Map.lookup tvarId tvars of
     Nothing                        -> error $ "Impossible: TVar not found: " ++ show tvarId
-    Just (TVarHandle _ _ tvarData) -> liftIO $ writeIORef tvarData $ encode a
+    Just (TVarHandle _ tvarData) -> liftIO $ writeIORef tvarData $ encode a
 
 interpretStmf :: STMF a -> Atomic (Either RetryCmd a)