+++ src/Control/Concurrent/STM/Free/TVar.hs
@@ -1,5 +1,5 @@
 module Control.Concurrent.STM.Free.TVar where
 
-import Control.Concurrent.STM.Free.Internal.Types
+import           Control.Concurrent.STM.Free.Internal.Types (TVarId)
 
 data TVar a = TVar TVarId