+++ ChangeLog.md
@@ -1,3 +1,10 @@
 # Changelog for stm-free
 
+0.1.0.0
+Initial version.
+
+- STML monadic language
+- Atomic evaluation of STML
+- Work with TVars containing ToJSON / FromJSON values
+
 ## Unreleased changes